feat(tools): brush size as an editable dropdown, decimals, up to 2000 (Paint.NET-identical)
Per the official Paint.NET docs, its brush size is an editable combo — type a
value or pick a preset — that accepts decimals (it antialiases sub-integer
sizes) and ranges to 2000. Ours was an integer spin box capped at 500 (#12).

Now:
- Tool::brushSize is a double, clamped to 1..2000 (was int, 1..500). The tools
  already did their radius/spacing maths in doubles, so drawing is unchanged
  besides honouring the fractional size.
- The tool-options control is an editable QComboBox with ~60 presets marching
  1..2000, a QDoubleValidator, and commit-on-Enter / on-select (not per
  keystroke, so "6.5" isn't cut short — same care as the resize dialog). The
  [ ] and Ctrl+[ Ctrl+] shortcuts still nudge by 1 and 5.

Tests: 284 -> 291. The new end-to-end check types "12.5" into the dropdown and
asserts the tool's size becomes 12.5, plus the 1..2000 clamp.

Note on the sibling request (#13, a 54-entry "brush texture" dropdown): the
Paint.NET docs list no such control — the paintbrush has Size, Hardness,
Spacing, Smoothing and a Fill Style (solid/pattern) selector, not a texture
library. Not implementing a feature the reference app doesn't have.

+++ src/tools/clonestamptool.cpp
@@ -58,7 +58,7 @@ void CloneStampTool::mouseReleaseEvent(const QPointF &, QMouseEvent *, CanvasWid
 }
 
 void CloneStampTool::cloneDab(const QPointF &pos, QImage &target, const QImage &source) {
-    int r = std::max(1, m_brushSize / 2);   // size 1 must not give r=0 (÷0 → NaN)
+    int r = std::max(1, static_cast<int>(m_brushSize / 2));   // size 1 must not give r=0 (÷0 → NaN)
     QPointF srcPos = pos + m_offset;
     const double hard = m_hardness / 100.0;
     const double strength = (m_opacity / 100.0) * m_pressure;
